ref(server): Cleans up ClientName and PlatformTag, adds Relay client name
