+++ relay-server/src/services/processor/event.rs
@@ -29,7 +29,7 @@ use crate::services::processor::{
     MINIMUM_CLOCK_DRIFT, ProcessingError, SpansExtracted, event_category, event_type,
 };
 use crate::services::projects::project::ProjectInfo;
-use crate::statsd::{PlatformTag, RelayCounters, RelayHistograms, RelayTimers};
+use crate::statsd::{RelayCounters, RelayHistograms, RelayTimers};
 use crate::utils::{self, ChunkedFormDataAggregator, FormDataIter};
 
 /// Result of the extraction of the primary event payload from an envelope.
@@ -204,11 +204,12 @@ pub fn finalize<Group: EventProcessing>(
         inner_event._metrics = Annotated::new(metrics);
 
         if inner_event.ty.value() == Some(&EventType::Transaction) {
+            let platform = utils::PlatformTag::from(inner_event.platform.as_str());
+
             metric!(
                 counter(RelayCounters::EventTransaction) += 1,
                 source = utils::transaction_source_tag(inner_event),
-                platform =
-                    PlatformTag::from(inner_event.platform.as_str().unwrap_or("other")).name(),
+                platform = platform.name(),
                 contains_slashes = if inner_event
                     .transaction
                     .as_str()
@@ -225,7 +226,7 @@ pub fn finalize<Group: EventProcessing>(
             metric!(
                 histogram(RelayHistograms::EventSpans) = span_count,
                 sdk = envelope.meta().client_name().name(),
-                platform = inner_event.platform.as_str().unwrap_or("other"),
+                platform = platform.name(),
             );
 
             let has_otel = inner_event
@@ -237,7 +238,7 @@ pub fn finalize<Group: EventProcessing>(
                 metric!(
                     counter(RelayCounters::OpenTelemetryEvent) += 1,
                     sdk = envelope.meta().client_name().name(),
-                    platform = inner_event.platform.as_str().unwrap_or("other"),
+                    platform = platform.name(),
                 );
             }
         }
