+++ relay-server/src/statsd.rs
@@ -977,150 +977,3 @@ impl CounterMetric for RelayCounters {
         }
     }
 }
-
-/// Low-cardinality platform that can be used as a statsd tag.
-pub enum PlatformTag {
-    Cocoa,
-    Csharp,
-    Edge,
-    Go,
-    Java,
-    Javascript,
-    Julia,
-    Native,
-    Node,
-    Objc,
-    Other,
-    Perl,
-    Php,
-    Python,
-    Ruby,
-    Swift,
-}
-
-impl PlatformTag {
-    pub fn name(&self) -> &str {
-        match self {
-            Self::Cocoa => "cocoa",
-            Self::Csharp => "csharp",
-            Self::Edge => "edge",
-            Self::Go => "go",
-            Self::Java => "java",
-            Self::Javascript => "javascript",
-            Self::Julia => "julia",
-            Self::Native => "native",
-            Self::Node => "node",
-            Self::Objc => "objc",
-            Self::Other => "other",
-            Self::Perl => "perl",
-            Self::Php => "php",
-            Self::Python => "python",
-            Self::Ruby => "ruby",
-            Self::Swift => "swift",
-        }
-    }
-}
-
-impl<S: AsRef<str>> From<S> for PlatformTag {
-    fn from(value: S) -> Self {
-        match value.as_ref() {
-            "cocoa" => Self::Cocoa,
-            "csharp" => Self::Csharp,
-            "edge" => Self::Edge,
-            "go" => Self::Go,
-            "java" => Self::Java,
-            "javascript" => Self::Javascript,
-            "julia" => Self::Julia,
-            "native" => Self::Native,
-            "node" => Self::Node,
-            "objc" => Self::Objc,
-            "perl" => Self::Perl,
-            "php" => Self::Php,
-            "python" => Self::Python,
-            "ruby" => Self::Ruby,
-            "swift" => Self::Swift,
-            _ => Self::Other,
-        }
-    }
-}
-
-/// Low-cardinality SDK name that can be used as a statsd tag.
-pub enum ClientName<'a> {
-    Ruby,
-    CocoaFlutter,
-    CocoaReactNative,
-    Cocoa,
-    Dotnet,
-    AndroidReactNative,
-    AndroidJava,
-    SpringBoot,
-    JavascriptBrowser,
-    Electron,
-    NestJs,
-    NextJs,
-    Node,
-    React,
-    Vue,
-    Native,
-    Laravel,
-    Symfony,
-    Php,
-    Python,
-    Other(&'a str),
-}
-
-impl ClientName<'_> {
-    pub fn name(&self) -> &'static str {
-        match self {
-            Self::Ruby => "sentry-ruby",
-            Self::CocoaFlutter => "sentry.cocoa.flutter",
-            Self::CocoaReactNative => "sentry.cocoa.react-native",
-            Self::Cocoa => "sentry.cocoa",
-            Self::Dotnet => "sentry.dotnet",
-            Self::AndroidReactNative => "sentry.java.android.react-native",
-            Self::AndroidJava => "sentry.java.android",
-            Self::SpringBoot => "sentry.java.spring-boot.jakarta",
-            Self::JavascriptBrowser => "sentry.javascript.browser",
-            Self::Electron => "sentry.javascript.electron",
-            Self::NestJs => "sentry.javascript.nestjs",
-            Self::NextJs => "sentry.javascript.nextjs",
-            Self::Node => "sentry.javascript.node",
-            Self::React => "sentry.javascript.react",
-            Self::Vue => "sentry.javascript.vue",
-            Self::Native => "sentry.native",
-            Self::Laravel => "sentry.php.laravel",
-            Self::Symfony => "sentry.php.symfony",
-            Self::Php => "sentry.php",
-            Self::Python => "sentry.python",
-            Self::Other(_) => "other",
-        }
-    }
-}
-
-impl<'a> From<&'a str> for ClientName<'a> {
-    fn from(value: &'a str) -> Self {
-        match value {
-            "sentry-ruby" => Self::Ruby,
-            "sentry.cocoa.flutter" => Self::CocoaFlutter,
-            "sentry.cocoa.react-native" => Self::CocoaReactNative,
-            "sentry.cocoa" => Self::Cocoa,
-            "sentry.dotnet" => Self::Dotnet,
-            "sentry.java.android.react-native" => Self::AndroidReactNative,
-            "sentry.java.android" => Self::AndroidJava,
-            "sentry.java.spring-boot.jakarta" => Self::SpringBoot,
-            "sentry.javascript.browser" => Self::JavascriptBrowser,
-            "sentry.javascript.electron" => Self::Electron,
-            "sentry.javascript.nestjs" => Self::NestJs,
-            "sentry.javascript.nextjs" => Self::NextJs,
-            "sentry.javascript.node" => Self::Node,
-            "sentry.javascript.react" => Self::React,
-            "sentry.javascript.vue" => Self::Vue,
-            "sentry.native" => Self::Native,
-            "sentry.php.laravel" => Self::Laravel,
-            "sentry.php.symfony" => Self::Symfony,
-            "sentry.php" => Self::Php,
-            "sentry.python" => Self::Python,
-            other => Self::Other(other),
-        }
-    }
-}
