+++ relay-server/src/utils/statsd.rs
@@ -77,3 +77,75 @@ where
 
     res
 }
+
+/// Low-cardinality platform that can be used as a statsd tag.
+pub enum PlatformTag {
+    Cocoa,
+    Csharp,
+    Edge,
+    Go,
+    Java,
+    Javascript,
+    Julia,
+    Native,
+    Node,
+    Objc,
+    Other,
+    Perl,
+    Php,
+    Python,
+    Ruby,
+    Swift,
+}
+
+impl PlatformTag {
+    pub fn name(&self) -> &str {
+        match self {
+            Self::Cocoa => "cocoa",
+            Self::Csharp => "csharp",
+            Self::Edge => "edge",
+            Self::Go => "go",
+            Self::Java => "java",
+            Self::Javascript => "javascript",
+            Self::Julia => "julia",
+            Self::Native => "native",
+            Self::Node => "node",
+            Self::Objc => "objc",
+            Self::Other => "other",
+            Self::Perl => "perl",
+            Self::Php => "php",
+            Self::Python => "python",
+            Self::Ruby => "ruby",
+            Self::Swift => "swift",
+        }
+    }
+}
+
+impl From<&str> for PlatformTag {
+    fn from(value: &str) -> Self {
+        match value {
+            "cocoa" => Self::Cocoa,
+            "csharp" => Self::Csharp,
+            "edge" => Self::Edge,
+            "go" => Self::Go,
+            "java" => Self::Java,
+            "javascript" => Self::Javascript,
+            "julia" => Self::Julia,
+            "native" => Self::Native,
+            "node" => Self::Node,
+            "objc" => Self::Objc,
+            "perl" => Self::Perl,
+            "php" => Self::Php,
+            "python" => Self::Python,
+            "ruby" => Self::Ruby,
+            "swift" => Self::Swift,
+            _ => Self::Other,
+        }
+    }
+}
+
+impl From<Option<&str>> for PlatformTag {
+    fn from(value: Option<&str>) -> Self {
+        value.map(Self::from).unwrap_or(PlatformTag::Other)
+    }
+}
